backend_note_taking_app/requirements.txt
Django
djangorestframework
django-filter
djangorestframework-simplejwt
drf_spectacular
django-cors-headers
drf_yasg